+++ spec/unit/honeybadger_spec.rb
@@ -132,7 +132,7 @@ describe Honeybadger do
 
     it "generates a backtrace excluding the singleton" do
       expect(instance.worker).to receive(:push) do |notice|
-        expect(notice.backtrace.to_a[0][:file]).to eq('[PROJECT_ROOT]/spec/unit/honeybadger_spec.rb')
+        expect(notice.backtrace.to_a[0]).to match('lib/honeybadger/agent.rb')
       end
 
       Honeybadger.notify(error_message: 'testing backtrace generation')